backend/requirements.txt
# Core dependencies for AI-powered blog generation
pydantic>=2.9.0
pydantic-ai>=0.0.14

# RSS feed parsing
feedparser>=6.0.11

# Google Gemini support
google-generativeai>=0.8.3

# DeepSeek and OpenAI-compatible models (optional)
openai>=1.54.0

# Async HTTP requests (for future API integrations)
aiohttp>=3.10.0
httpx>=0.27.0

# Date/time utilities
python-dateutil>=2.9.0

# Environment variables
python-dotenv>=1.0.0

# Logging and utilities
colorlog>=6.8.0
